feat(theme): add useAppearance hook with design-language root attributes
