+++ src/renderer/components/GlobalEffects.tsx
@@ -1,10 +1,9 @@
 import { type FC, useEffect } from 'react';
 
-import { useTheme } from '@primer/react';
-
 import { useQueryClient } from '@tanstack/react-query';
 
 import { useAccounts } from '../hooks/useAccounts';
+import { useAppearance } from '../hooks/useAppearance';
 import { useNotifications } from '../hooks/useNotifications';
 import { useOnlineStatus } from '../hooks/useOnlineStatus';
 import {
@@ -14,14 +13,6 @@ import {
 import { useAccountsStore, useFiltersStore, useSettingsStore } from '../stores';
 
 import { setTrayIconColorAndTitle } from '../utils/system/tray';
-import {
-  DEFAULT_DAY_COLOR_SCHEME,
-  DEFAULT_DAY_HIGH_CONTRAST_COLOR_SCHEME,
-  DEFAULT_NIGHT_COLOR_SCHEME,
-  DEFAULT_NIGHT_HIGH_CONTRAST_COLOR_SCHEME,
-  mapThemeModeToColorMode,
-  mapThemeModeToColorScheme,
-} from '../utils/ui/theme';
 
 /**
  * Hosts all app-level side effects. Mount exactly once, inside the Primer
@@ -42,13 +33,10 @@ export const GlobalEffects: FC = () => {
   const useUnreadActiveIcon = useSettingsStore((s) => s.useUnreadActiveIcon);
   const useAlternateIdleIcon = useSettingsStore((s) => s.useAlternateIdleIcon);
 
-  // Subscribe to theme related settings for useEffect dependencies
-  const theme = useSettingsStore((s) => s.theme);
-  const increaseContrast = useSettingsStore((s) => s.increaseContrast);
-
-  const { setColorMode, setDayScheme, setNightScheme } = useTheme();
-
   const isOnline = useOnlineStatus();
+
+  // Appearance side effects: color mode/scheme + design-language/material attributes
+  useAppearance();
 
   // Notification fetching side effects (sound / native alerts, inactivity refetch)
   const { status, notificationCount } = useNotifications({ withSideEffects: true });
@@ -58,26 +46,6 @@ export const GlobalEffects: FC = () => {
 
   // Global keyboard shortcut registration, reverting on failure
   useShortcutRegistration();
-
-  // Theme
-  useEffect(() => {
-    const colorMode = mapThemeModeToColorMode(theme);
-    const colorScheme = mapThemeModeToColorScheme(theme, increaseContrast);
-
-    setColorMode(colorMode);
-
-    // When colorScheme is null (System theme), use appropriate fallbacks
-    // based on whether high contrast is enabled
-    const dayFallback = increaseContrast
-      ? DEFAULT_DAY_HIGH_CONTRAST_COLOR_SCHEME
-      : DEFAULT_DAY_COLOR_SCHEME;
-    const nightFallback = increaseContrast
-      ? DEFAULT_NIGHT_HIGH_CONTRAST_COLOR_SCHEME
-      : DEFAULT_NIGHT_COLOR_SCHEME;
-
-    setDayScheme(colorScheme ?? dayFallback);
-    setNightScheme(colorScheme ?? nightFallback);
-  }, [theme, increaseContrast, setColorMode, setDayScheme, setNightScheme]);
 
   // oxlint-disable-next-line react/exhaustive-deps -- We want to update the tray on setting or notification changes
   useEffect(() => {
